+++ build.gradle
@@ -45,39 +45,32 @@ subprojects {
 
 	dependencies {
 		compile ('com.aol.cyclops:cyclops-all:5.1.0')
-		compile 'com.nurkiewicz.lazyseq:lazyseq:0.0.1'
-		compile 'org.springframework.data:spring-data-jpa:1.7.2.RELEASE'
-		compile 'com.googlecode.genericdao:dao-hibernate:1.2.0'
-		compile group: 'org.hibernate', name:'hibernate-validator', version:'5.1.3.Final'
-		compile group: 'org.hibernate', name: 'hibernate-core', version:'4.3.8.Final'
-		compile group: 'org.hibernate', name: 'hibernate-entitymanager', version:'4.3.8.Final'
+		
 
 		compile ('com.ryantenney.metrics:metrics-spring:3.0.3'){
 			exclude(module: 'org.springframework')
 		}
 	
 		compile group: 'com.aol.simplereact', name:'simple-react', version:'0.98'
+		compile group: 'org.aspectj', name: 'aspectjrt', version:'1.5.4'
+		compile group: 'org.aspectj', name: 'aspectjweaver', version:'1.5.4'
+		
 		compile group: 'org.apache.httpcomponents', name:'httpclient', version:'4.4'
 		compile group: 'org.apache.httpcomponents', name:'httpasyncclient', version:'4.1-beta1'
+		
 		compile group: 'com.wordnik', name: 'swagger-jersey2-jaxrs_2.10', version:'1.3.10'
-		compile group: 'org.springframework', name: 'spring-web', version:'4.1.5.RELEASE'
-		compile group: 'org.springframework', name: 'spring-context', version:'4.1.5.RELEASE'
-		compile group: 'org.springframework', name: 'spring-context-support', version:'4.1.5.RELEASE'
+		compile group: 'org.springframework', name: 'spring-web', version:"${springVersion}"
+		compile group: 'org.springframework', name: 'spring-context', version:"${springVersion}"
+		compile group: 'org.springframework', name: 'spring-context-support', version:"${springVersion}"
 		compile group: 'ch.qos.logback', name: 'logback-classic', version:'1.0.4'
 		compile group: 'ch.qos.logback', name: 'logback-core', version:'1.0.4'
 		compile group: 'org.slf4j', name: 'jcl-over-slf4j', version:'1.6.1'
-		compile group: 'org.apache.commons', name: 'commons-math', version:'2.2'
+	
+
 		compile group: 'com.google.guava', name: 'guava', version:'18.0'
-		compile group: 'commons-lang', name: 'commons-lang', version:'2.6'
-		compile group: 'org.apache.commons', name: 'commons-io', version:'1.3.2'
-		compile group: 'org.apache.commons', name: 'commons-compress', version:'1.1'
+	
+
 		compile group: 'org.apache.commons', name: 'commons-dbcp2', version:'2.1'
-
-		compile group: 'org.springframework', name: 'spring-jdbc', version:'4.1.5.RELEASE'
-		compile group: 'org.springframework', name: 'spring-orm', version:'4.1.5.RELEASE'
-
-		compile group: 'org.codehaus.groovy', name: 'groovy-all', version:'2.3.3'
-		compile group: 'commons-collections', name: 'commons-collections', version:'3.2.1'
 		compile group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-guava', version:'2.5.2'
 		compile group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jdk8', version:'2.5.2'
 		compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version:'2.5.2'
@@ -96,29 +89,24 @@ subprojects {
 			exclude(module: 'jackson-mapper-asl')
 		}
 		compile group: 'javax.ws.rs', name: 'javax.ws.rs-api', version:'2.0'
-		compile group: 'org.aspectj', name: 'aspectjrt', version:'1.5.4'
-		compile group: 'org.aspectj', name: 'aspectjweaver', version:'1.5.4'
+
 		compile group: 'org.glassfish.grizzly', name: 'grizzly-http-server', version:"$grizzlyVersion"
 		compile group: 'org.glassfish.grizzly', name: 'grizzly-http-all', version:"$grizzlyVersion"
 		compile group: 'org.glassfish.jersey.containers', name: 'jersey-container-grizzly2-http', version:"$jerseyVersion"
 		compile group: 'org.glassfish.jersey.containers', name: 'jersey-container-grizzly2-servlet', version:"$jerseyVersion"
-		compile 'org.glassfish.jersey.media:jersey-media-multipart:2.17'
+		compile "org.glassfish.jersey.media:jersey-media-multipart:$jerseyVersion"
 		
 		compile "org.glassfish.jersey.ext:jersey-spring3:$jerseyVersion"
 		compile "org.glassfish.jersey.ext:jersey-bean-validation:$jerseyVersion"
-	
-		compile ('com.nurkiewicz.jdbcrepository:jdbcrepository:0.4.1'){
-			exclude(module: 'org.springframework')
-		}
-		compile 'commons-beanutils:commons-beanutils:1.9.2'
+	//	compile 'commons-beanutils:commons-beanutils:1.9.2'
 		compile 'javax.interceptor:javax.interceptor-api:1.2-b04'
 		compile 'javax.el:javax.el-api:2.2.4'
 		compile 'org.glassfish.web:javax.el:2.2.4'
-		testCompile group: 'com.cyrusinnovation', name: 'mockito-groovy-support', version:'1.3'
-		testCompile group: 'junit', name: 'junit', version:'4.12'
-		testCompile group: 'org.mockito', name: 'mockito-all', version:'1.9.5'
+
+		testCompile group: 'junit', name: 'junit-dep', version:'4.11'
+		testCompile group: 'org.mockito', name: 'mockito-core', version:'1.9.5'
 	
-		testCompile group: 'org.hsqldb', name:'hsqldb', version:'2.0.0'
+		
 
 		compile(group: 'org.projectlombok', name: 'lombok', version:'1.14.2') {
 			/* This dependency was originally in the Maven provided scope, but the project was not of type war.
