+++ gradle.properties
@@ -1,1 +1,2 @@
 version=0.59
+springVersion=4.1.5.RELEASE
