+++ micro-boot/src/main/java/com/aol/micro/server/spring/boot/JerseyApplication.java
@@ -2,11 +2,9 @@ package com.aol.micro.server.spring.boot;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Properties;
 
-import org.hibernate.SessionFactory;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.embedded.EmbeddedServletContainer;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerException;
@@ -16,13 +14,9 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 
 import com.aol.micro.server.module.Environment;
-import com.aol.micro.server.module.ModuleBean;
 import com.aol.micro.server.servers.AccessLogLocationBean;
-import com.aol.micro.server.spring.datasource.hibernate.DAOBuilder;
-import com.aol.micro.server.spring.datasource.hibernate.DAOProvider;
 import com.aol.micro.server.spring.properties.PropertyFileConfig;
 
 
@@ -69,13 +63,7 @@ public class JerseyApplication extends SpringBootServletInitializer {
 
 
 
-	public DAOProvider buildDAOProvider(
-			AnnotationConfigWebApplicationContext rootContext,
-			SessionFactory sessionFactory) {
-
-		return DAOBuilder.builder().applicationContext(rootContext)
-				.factory(sessionFactory).build().daoProvider();
-	}
+	
 
 	
 	@Bean
