+++ micro-core/src/main/java/com/aol/micro/server/config/Classes.java
@@ -11,12 +11,6 @@ import com.aol.micro.server.rest.resources.ConfigureResources;
 import com.aol.micro.server.servers.AccessLogConfig;
 import com.aol.micro.server.spring.datasource.DataSourceBuilder;
 import com.aol.micro.server.spring.datasource.JdbcConfig;
-import com.aol.micro.server.spring.datasource.hibernate.DAOProvider;
-import com.aol.micro.server.spring.datasource.hibernate.GenericHibernateService;
-import com.aol.micro.server.spring.datasource.hibernate.HibernateConfig;
-import com.aol.micro.server.spring.datasource.hibernate.SpringDataConfig;
-import com.aol.micro.server.spring.datasource.jdbc.RomaRowMapperConfig;
-import com.aol.micro.server.spring.datasource.jdbc.SQL;
 import com.aol.micro.server.spring.metrics.CodahaleMetricsConfigurer;
 import com.aol.micro.server.spring.properties.PropertyFileConfig;
 
@@ -38,12 +32,8 @@ public enum Classes {
 	 */
 	CORE_CLASSES(PropertyFileConfig.class, MiscellaneousConfig.class, AopConfig.class, CodahaleMetricsConfigurer.class,
 			ConfigureActiveJobsAspect.class, ScheduleAndAsyncConfig.class, ConfigureResources.class, ConfigureEnviroment.class, AccessLogConfig.class),
-	JDBC_CLASSES(JdbcConfig.class, DAOProvider.class, DataSourceBuilder.class, SQL.class, SpringDataConfig.class),
-	ROMA_ROW_MAPPER(RomaRowMapperConfig.class),
-	HIBERNATE_CLASSES(HibernateConfig.class, JdbcConfig.class, GenericHibernateService.class, DAOProvider.class, DataSourceBuilder.class, SQL.class),
-	SPRING_DATA_CLASSES(SpringDataConfig.class),
-	DATASOURCE_CLASSES(JdbcConfig.class, DataSourceBuilder.class);
-
+	DATA_SOURCE_CLASSES(JdbcConfig.class,DataSourceBuilder.class);
+	
 	@Getter
 	private final Class[] classes;
