--- micro-core/src/main/java/com/aol/micro/server/spring/datasource/jdbc/RomaRowMapperConfig.java
@@ -1,10 +1,0 @@
-package com.aol.micro.server.spring.datasource.jdbc;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.ImportResource;
-
-@Configuration
-@ImportResource("classpath*:roma-context.xml")
-public class RomaRowMapperConfig {
-
-}
