+++ micro-core/src/main/java/com/aol/micro/server/utility/UsefulStaticMethods.java
@@ -3,29 +3,6 @@ package com.aol.micro.server.utility;
 
 import static java.util.Optional.ofNullable;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
-import java.util.TreeSet;
-
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.io.FileUtils;
-import org.codehaus.groovy.runtime.DefaultGroovyMethods;
-import org.springframework.aop.framework.Advised;
-import org.springframework.aop.support.AopUtils;
-
-import com.google.common.base.Function;
-import com.google.common.collect.ImmutableSortedSet;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Multimap;
-import com.google.common.collect.TreeMultimap;
-
 public class UsefulStaticMethods {
