--- micro-core/src/main/java/com/googlecode/genericdao/dao/hibernate/GenericDAOAccessor.java
@@ -1,9 +1,0 @@
-package com.googlecode.genericdao.dao.hibernate;
-
-public class GenericDAOAccessor {
-
-	
-	public static void setTargetType(GenericDAOImpl dao,Class targetType){
-		dao.persistentClass =targetType;
-	}
-}
