--- micro-core/src/test/java/app/spring/data/jpa/com/aol/micro/server/HibernateEntityRepository.java
@@ -1,9 +1,0 @@
-package app.spring.data.jpa.com.aol.micro.server;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-
-public interface HibernateEntityRepository extends CrudRepository<HibernateEntity, Long> {
-
-}
