+++ micro-data/src/main/java/com/aol/micro/server/spring/datasource/hibernate/SpringDataConfig.java
