+++ micro-data/src/test/java/app/jdbc/com/aol/micro/server/HibernateEntityForDDLAutoCreate.java
